requirements-airflow.txt
apache-airflow-providers-apache-spark==4.2.0
pyspark==3.3.2
pandas==1.5.3
numpy==1.23.5
apache-airflow-providers-amazon==3.4.0
apache-airflow-providers-redis==3.3.0